+++ projects/aca-content/src/lib/aca-content.routes.ts
@@ -40,7 +40,7 @@ import { Route } from '@angular/router';
 import { SharedLinkViewComponent } from './components/shared-link-view/shared-link-view.component';
 import { TrashcanComponent } from './components/trashcan/trashcan.component';
 import { ShellLayoutComponent } from '@alfresco/adf-core/shell';
-import { SearchAiResultsComponent } from './components/ai/search-ai/search-ai-results/search-ai-results.component';
+import { SearchAiResultsComponent } from './components/knowledge-retrieval/search-ai/search-ai-results/search-ai-results.component';
 
 export const CONTENT_ROUTES: ExtensionRoute[] = [
   {
