--- projects/aca-content/src/lib/components/ai/search-ai/toggle-search/toggle-search.component.scss
@@ -1,16 +1,0 @@
-// stylelint-disable selector-class-pattern
-
-.aca-toggle-search {
-  .mat-slide-toggle.mat-checked .mat-slide-toggle-thumb {
-    background-color: white;
-  }
-
-  .mat-slide-toggle.mat-checked .mat-slide-toggle-bar {
-    background: linear-gradient(to right, #2a7de1, #902ae1);
-  }
-
-  .mat-slide-toggle-content {
-    font-weight: 500;
-    font-size: 20px;
-  }
-}
