+++ projects/aca-content/src/lib/components/favorites/favorites.component.ts
@@ -40,8 +40,8 @@ import { DocumentListModule } from '@alfresco/adf-content-services';
 import { DataTableModule, EmptyContentComponent, PaginationComponent } from '@alfresco/adf-core';
 import { DocumentListDirective } from '../../directives/document-list.directive';
 import { TranslateModule } from '@ngx-translate/core';
-import { SearchInputContainerComponent } from '../ai/search-ai/search-input-container/search-input-container.component';
 import { SearchAIService } from '../../services/search-ai.service';
+import { SearchAiInputContainerComponent } from '../knowledge-retrieval/search-ai/search-ai-input-container/search-ai-input-container.component';
 
 @Component({
   standalone: true,
@@ -57,7 +57,7 @@ import { SearchAIService } from '../../services/search-ai.service';
     PageLayoutComponent,
     TranslateModule,
     ToolbarComponent,
-    SearchInputContainerComponent,
+    SearchAiInputContainerComponent,
     EmptyContentComponent,
     DynamicColumnComponent
   ],
