+++ projects/aca-content/src/lib/components/files/files.component.html
@@ -1,9 +1,9 @@
 <aca-page-layout [hasError]="!isValidPath">
   <div class="aca-page-layout-header">
-    <aca-search-input-container
+    <aca-search-ai-input-container
       @fadeInFromLeft
       *ngIf="showAISearchInput; else breadcrumbTemplate">
-    </aca-search-input-container>
+    </aca-search-ai-input-container>
     <ng-template #breadcrumbTemplate>
       <div class="aca-header-container">
         <adf-breadcrumb [root]="title"
