+++ projects/aca-content/src/lib/components/knowledge-retrieval/search-ai/search-ai-input-container/search-ai-input-container.component.html
@@ -1,4 +1,4 @@
-<aca-search-ai-input [hideAIToggle]="true" (searchSubmitted)="onAIInputSearchSubmitted()" [placeholder]="placeholder"></aca-search-ai-input>
+<aca-search-ai-input (searchSubmitted)="onAIInputSearchSubmitted()" [placeholder]="placeholder"></aca-search-ai-input>
 <mat-divider [vertical]="true" class="aca-search-input-container-divider"></mat-divider>
 <button mat-icon-button (click)="leaveSearchInput()">
   <mat-icon>close</mat-icon>
