+++ projects/aca-content/src/lib/components/knowledge-retrieval/search-ai/search-ai-input-container/search-ai-input-container.component.scss
@@ -1,4 +1,4 @@
-aca-search-input-container {
+aca-search-ai-input-container {
   display: flex;
   flex-direction: row;
   flex: 1;
