+++ projects/aca-content/src/lib/components/knowledge-retrieval/search-ai/search-ai-results/search-ai-results.component.html
@@ -1,6 +1,6 @@
 <aca-page-layout @fadeIn>
     <div class="aca-page-layout-header">
-        <aca-search-input-container [placeholder]="'Would you like to ask anything else?'"></aca-search-input-container>
+        <aca-search-ai-input-container [placeholder]="'Would you like to ask anything else?'"></aca-search-ai-input-container>
     </div>
 
     <div class="aca-page-layout-content">
