+++ projects/aca-content/src/lib/components/knowledge-retrieval/search-icon/search-icon.component.ts
